Inventory loading from JSON
This is a basic implementation of loading an inventory from a JSON file. This could be useful for changing the inventory quickly without needing to recompile.

An inventory can be loaded by making a inventory.json file in the Win64 folder (the directory where Fortnite executable is stored) and making an array called items. If no file is present then the original inventory will be loaded.